Create certainty.py
Tried to create functions that can create a certainty map from an image. Ended up pretty bad and needs more work. Still got some results to show for it despite the poor quality of the actual results.

In the inpainting_with_certainty.py file we combine all the results from the certainty variants created and then use them to try to reconstruct the image. Note however that we try to do this on a complete image so I'll have to create an image which lacks information before evaluating the actual results (even though the results from this indicates that my current confidence mapping method is lacking since it distorts the image).